requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2>=2.10.1
MarkupSafe==1.0
Werkzeug==0.15.5

pydicom
scikit-image
#opencv-python
azure-storage-blob==12.3.2
pillow
pymongo
flask_pymongo
python-dotenv
dnspython
